Project Documentation
Finished up to the design stage, made a choice of the NASA API and currently working on Gantt Chart. Have kept the test code for the NASA API the same, where you can view the current APOD in a text based system. However, will transition to a GUI based system.