Fixed some of the discrepancies at home.
So I noted at school that the button would not open the image in the web browser. I later found out that this was due to the try and except functions, where it would first try the code for printing and opening the URL and furthermore detecting an error. This, for some reason posed an error, especially with the fact that I used the lambda function in the button when I didn't need it, and I also packed the command in the same line possibly making it not work. However, I made my code much more simpler to utilise and have created a basic button to open the APOD in a web browser.