Did a LOT of work on the GUI.
Today I created a GUI which allows you to view the Daily APOD inside as well as the chosen date of an APOD from the past. This allows for a lot more flexibility around the APOD database. I added parameters for date in the my_module.py in order to call the date in the GUI, and also created individual frames and labels for the NASA APOD's to be viewed in the GUI. My next goal would be possibly showing the information of the APOD which would require text labels and new parameters of the title, explanation and the url of the APOD. Further on, I will possibly implement a menu to save certain APOD's that people would like and to remove these APOD's. This GUI is mainly for the use of people who are interested in APOD's so far, and I may implement other aspects such as matplotlib and pandas if features are present in the NASA api.