Finally made the explanation, title, url and date viewable in the GUI along with the image.
So I made a text label that derives the explanation, title, url, and date. Basically how this works is from the original apod =  mm.get_apod() function. I kept the module and the function inside the same, and derived the following explanation, date, title, and url and reformatted it to fit in a text widget, making it a simple task. I also created some visual changes to the text widget, adding white font and black background. This is only a staple for now, as I can customise these styles in the future, as shown in the implication of the style class to make it more space themed. As for now, these are my accomplishments, and my goal remains to create a library where one can save these APOD's and remove them from that library. I can create visual changes for a more space themed GUI, and I also wish to utilise a calender in my GUI for more easy access. If there is a module, I will use it if time is sensible.