Calendar Function MADE!
I made a calendar function for full user optimisation in my program, specifically for viewing APOD's by date. I did this through the tkcalendar, date entry widget, and I also used a formatting tool for the date entry called ttkbootstrap, allowing me to make the calendar look better in general. My goal now is to create an introduction to the GUI, so users know how to utilise the application.