Today I found out something new
Today a new error occurred, which was a key error. This occurred due to an APOD video instead of an image, which I was able to fix with a new exception in the module. Also, I added a button in the FindFrame for a daily apod that does this, but I havent done so for the viewframe.